refactor: add session to some message types
Each frontend should be associated with a session that uniquely identify it in the backend. We are not currently using them in the backend but it will prevent `event_message.tab` collisions in the future. `session` is added to the message types: `event`, `amend`, `group`, and `error`.

`session` is added to the following message