refactor: reduce the scope of stop-all messages
When a stop-all message comes from a socket, we want to stop only the tracks associated to that socket. To do that we maintain a set of active tracks per connection. We also use that set of tracks to close them all once the socket disconnects. NB: `hook-exit` sends stop-all message upon exiting the process but this message never arrive, so this serves as a fallback mechanism.