refactor: tweak log level of source map
When the transpiler adds a function that is not in the source, the source map will have a missing segment/group. Hence, it is expected behavior and should not be reported to the user.

However, when the source map returns an out-of-range source index, there is something wrong with the source map and it should be reported as a warning.