fix: more robust ast lookup
Sometimes source maps provide inaccurate source code position. To solve this issue, we implemented a recursive lookup algorithm lenient on position matching. But this algorithm would not handle the following case: `export default function f () {};` where the mapped source position was `{line:1, column:0}` (the accurate position is `{line:1, column:15}`). Since recovering source map inaccuracies was already a problem in the past, I decided to invest on a bullet-proof algorithm: functions are keyed with their position. If an exact match does not exists, we pick the next best match which is based on computing position distance. This distance emphases lines distance over columns distance.