fix: Support class features when transforming source code
appmap-agent-js uses @babel/parser to parse the ECMAScript code for
transformation; the output generated by that library is not 100%
estree-compliant [1]. It has an `estree` plugin that reverts the
deviations; however, by default not all of them are supported
(apparently for @babel/parser's own backward compatibility reasons).
Fortunately it has a config setting that enables them.

One of these differences is that the standard node type of
PropertyDefinition is called ClassProperty instead, which made the
transformer unable to correctly handle ES classes.

[1] https://babeljs.io/docs/babel-parser#output

Fixes #199