refactor: add an assertion to help diagnose invalid callstack state
Bugs related to invalid callstack states are the hardest to diagnose because they are detected during trace post-processing long past their root cause. This assertion should help diagnosing the ones that are related to generator functions and asynchronous functions. To be compact, failure is signaled by reading a missing variable.