perf: frontend no longer validate data
Important validation (mostly user-provided configuration) happens on the backend. Note that `recorder-cli` can still use the `validate` component but it is a no-op. That is because we still want to perform validation in testing.

This is the first part of an effort to reduce exteral dependencies to enable browser support.