feat: can record files with external urls
Previously, only files whose urls could be expressed relatively to the url of the project (which was always a file url) could be represented in the appmap. Now we are able to include files of any url. The relative representation if it exists is still preferred. If the relative representation does not exists, the origin of the external url will be displayed as a top-level package.

I moved some pure functions in `components/classmap/default/loc.mjs` and `components/classmpa/default/specifier.mjs` to ease testing.

There is a small breaking change which does not warrant a major version bump imo: the relative paths in the appmaps are always start with `./` or `../`. This is to match node module specifier.