refactor: change service wrapper into pool of sockets
The main responsibility of the `service` component was to keep track of the connections of a `net.Server` in order to close them all when it needs to be interrupted. Before, this was achieved by wrapping `net.Server` into something called `service`. This added an unnecessary abstraction layer. We can provide the same functionality as a pool of sockets which does not force the api consumer to use promises. Later we will reuse the `pool` component inside the man-in-the-middle proxy.