refactor: move parse in dedicated component
This is the first step to get rid of the `source` component. It was a confusing abstraction because it was passed around pure and impure parts of the agents. Plus, it was weak as it only provides caching of hash and estree. We can do that better in the consumers of `source`.