refactor: update instrumentation to consume the new source component
The new source component represent a single source file whereas the old classmap component represents a set of source files. Managements of multiple source files is done in `codebase.mjs`.

Also, I changed to API of the instrumentation component to enable a cleaner separation between pure and impure code. The two methods `extractMissingUrlArray` and `instrument` are given `cache` which is a mapping from `url` to `content. The consumer of this API is responsible to load files. This enables to move the components `mapping` and `mapping-file` in instrumentation.