refactor: rewrite classmap component into source component
The classmap component became needlessly complicated. I tried to simplify it iteratively but ended up completely rewriting it :(.

The classmap component was shared across the frontend to apply exclusion and backend to: (re)apply exclusion, lookup closure information by location and export classmap data.

The new source component now only represents a single source file which simplifies its usage on the frontend.

This commits does not build and is logically linked to the following two next commits. They update the consumers of the old `classmap` component: `instrumentation` and `trace`. This was made to help reviewing.