refactor: listen to exit event on browser
In the browser, we attempt to record a stop track event when 
`onbeforeunload` is emitted on `window`. This is to bring parity with 
the node behavior. Note that the backend will be notified regardless 
when the websocket disconnect. In the future, this may be used to add 
information on the exit event to the appmap.