refactor: buffer messages in the frontend rather than in the socket
This refactoring is to move more logic on the frontend component which 
is pure (easier to test and maintain). This enables getting rid of the 
agent and emitter abstractions and directly let the recorders manipulate 
message producers (hooks) and message consumers (flush). Also, sockets 
are simpler and no longer have to implement buffering separately.

In the future, this refactoring will enable implementing an off/on 
toggle to prevent infinite recursion on calling `toString`. It will also 
enable performance improvement by compressing batch of messages before 
sending them to the network.