feat: Crash reports are written to disk on exit
This is crucial because we cannot run any asynchronous processes such as
emitting crash reports upon receiving an unhandled exception. By that
point, we're only able to perform synchronous logic. As a result, any
exceptions that have yet to be emitted must be queued and persisted to
be emitted on the next run.