fix: solve oom issue on large test suite
The backend does not process data as fast as the frontends generate it. 
So data buildups in the backend process and causes it to run out of 
memory. To solve this issue, I implemented back pressure. This required 
to implement two-way communication between the backend and the frontend. 
Now, the backend tells frontends how many traces it still needs to 
process and the frontends wait a proportional amount of time before 
starting a new test case.

This back pressure system has only be implemented for the mocha and jest 
recorder. Implementing it for the browser and node recorders would 
entail to synchronously block the main js thread which is a big no-no.