refactor: default to node builtin socket instead of posix-socket
The `posix-socket` module provides fast synchronous communication. This 
is important to make sure that data is actually sent to the backend. For 
instance, the `exit` event of `process` is synchronous so it is hard to 
guarantee that the socket is flushed before the process truly exit.

But with the two-way communication required to implement back pressure, 
it became slow due to the new pulling mechanism. Let's try to default to 
the builtin net socket. If things goes well, it would be warranted to 
get rid of the `posix-socket` dependency altogether.