refactor: make top-level record function async
For the recorders process and remote on node it makes sense to wait for 
the socket to open before proceeding to record the observed application. 
That is because we can't schedule async operations on exit. And if the 
application is synchronous is it hard to make sure data is actually 
sent.