fix: Expand short paths on Windows
Sometimes paths on Windows are represented by short paths such
as C:\PROGRA~1. This leads to issues if we try to compare paths
and one of them is equivalent to the other, but shortened.

To solve this issue, try to resolve the long path when resolving
directory paths (which is the only place I've seen this problem
occur so far). It specifically lead to failures on CI where TEMP
directory is a short path.